requirements.txt
django
python-dotenv==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
django-cors-headers==4.3.1